JL8-27 refactor : BottomSheet 공용 컴포넌트 리펙토링
